+++ lib/data.js
@@ -1,0 +1,4 @@
+export const data = {
+    aboutDescription: `My name is Ahsan Awadullah Azizan, and people often call me Ahsan. I'm currently living in Malang, but I originally lived in Tangerang Selatan. I'm a student at SMK Telkom Malang, majoring in Software Engineering. My hobbies includes coding, gaming, and football.`,
+
+}
